create dummy Ed25519Facade
this can later provide a real implementation but for now we can depend on it and mock it away

#tutadb2078
