+++ packages/tutanota-crypto/lib/index.ts
@@ -23,7 +23,6 @@ export {
 	x25519Encapsulate,
 	x25519Decapsulate,
 } from "./encryption/X25519.js"
-export { Ed25519PrivateKey, Ed25519PublicKey, Ed25519KeyPair } from "./encryption/Ed25519.js"
 export { generateRandomSalt, generateKeyFromPassphrase as generateKeyFromPassphraseBcrypt } from "./hashes/Bcrypt.js"
 export {
 	LibOQSExports,
