+++ src/common/api/worker/facades/KeyAuthenticationFacade.ts
@@ -1,8 +1,9 @@
 import { CryptoWrapper } from "../crypto/CryptoWrapper.js"
 import { concat, KeyVersion } from "@tutao/tutanota-utils"
-import { Aes256Key, AesKey, bitArrayToUint8Array, Ed25519PublicKey, MacTag, PQPublicKeys, X25519PublicKey } from "@tutao/tutanota-crypto"
+import { Aes256Key, AesKey, bitArrayToUint8Array, MacTag, PQPublicKeys } from "@tutao/tutanota-crypto"
 import { assertWorkerOrNode } from "../../common/Env.js"
 import { KeyMac } from "../../entities/sys/TypeRefs.js"
+import { Ed25519PublicKey } from "./Ed25519Facade"
 
 assertWorkerOrNode()
